projeto-individual-2/felipe-amorim-de-araujo/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astropy==7.2.0
astropy-iers-data==0.2026.4.6.0.54.57
astroquery==0.4.11
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
fastapi==0.135.3
filelock==3.25.2
flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface-hub==1.10.1
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.1.2
jaraco-functools==4.4.0
jeepney==0.9.0
jinja2==3.1.6
joblib==1.5.3
keyring==25.7.0
kiwisolver==1.5.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
pluggy==1.6.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pyerfa==2.0.1.5
pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyvo==1.8.1
pyyaml==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
secretstorage==3.5.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.3
soupsieve==2.8.3
sqlalchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.3
triton==3.6.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
webencodings==0.5.1
werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.0
